prevent infinite wait if tx has been dropped or does not exists
TXs can be dropped either because they are expired or the solver could not solve them. In that instance, we do not have to wait because all we do while waiting is to fetch the status of the tx on chain. If tx was dropped, then it cannot be onchain