[ BDH-235 ] Bundler v2 updates (#2)
* fix CTRL+C not actually closing the process

* fix otel collector. Set default srvice name and create new config

* use config value to determine wait timeout

* use context timeout cancellation as time#After does not garbage collect

* remove unnneeded repetitions

* create rpc middleware file and add asic test

* move main to cmd package and have rpc package as an actual server

* move otel logic to rpc server

* add sanity checks for http calls

* fix review and make conversion extremely safe

* prevent infinite wait if tx has been dropped or does not exists

TXs can be dropped either because they are expired or the solver could not solve them. In that instance, we do not have to wait because all we do while waiting is to fetch the status of the tx on chain. If tx was dropped, then it cannot be onchain